refactor: extract release-channel cache module with behavior parity (#671)
